feat: replace ConfigureConverterModal with ConfigureToolModal and refactor converter configuration steps
